[issue-24] Renamed a couple files and created dummy FXML files
