+++ src/main/java/io/github/palprz/view/translation/TranslationWindowController.java
@@ -1,0 +1,6 @@
+package io.github.palprz.view.translation;
+
+
+public class TranslationWindowController {
+
+}
